Sequence generation based on range of input.

Generates regular sequences from the minimum to the maximum of the input.

seq_range(x, length.out, extend = 0, procent = FALSE)

Arguments

x
Vector whoes range is used to generate the sequence.

length.out
desired length of the sequence. A non-negative number.

extend
Numeric, number of which to extend the range by.

procent
Indicates if extend is taken as a procentage.

Value

returns a vector of type "integer" or "double".

Details

The extend parameter can be either of length 1 or 2. If the length is 1 then the number is subtracted from the minimum, and added to the maximum. Thus extending the range by extend * 2. If the length is 2 then the first number is subtracted from the minimum and the second number is added to the maximum.

If procent is set to TRUE then extend is taken as a procentage where 1 is equal to the range, i.e. the distance from the minimum to the maximum.

Examples

 seq_range(1:5, 10)

#>  [1] 1.000000 1.444444 1.888889 2.333333 2.777778 3.222222 3.666667 4.111111
#>  [9] 4.555556 5.000000

 seq_range(1:5, 10, 1)

#>  [1] 0.0000000 0.6666667 1.3333333 2.0000000 2.6666667 3.3333333 4.0000000
#>  [8] 4.6666667 5.3333333 6.0000000

 seq_range(1:5, 10, 1:2)

#>  [1] 0.0000000 0.7777778 1.5555556 2.3333333 3.1111111 3.8888889 4.6666667
#>  [8] 5.4444444 6.2222222 7.0000000

 seq_range(rnorm(100), 10)

#>  [1] -1.9580910 -1.5116980 -1.0653050 -0.6189120 -0.1725191  0.2738739
#>  [7]  0.7202669  1.1666599  1.6130529  2.0594459

 seq_range(rnorm(100), 10, 0.1, TRUE)

#>  [1] -2.8332570 -2.2373226 -1.6413883 -1.0454539 -0.4495196  0.1464148
#>  [7]  0.7423491  1.3382835  1.9342179  2.5301522

 seq_range(1:5, 10, extend = 1, procent = TRUE)

#>  [1] -3.0000000 -1.6666667 -0.3333333  1.0000000  2.3333333  3.6666667
#>  [7]  5.0000000  6.3333333  7.6666667  9.0000000